# docs/docs/swagger.yaml
basePath: /
definitions:
...
  internal_transport_http.Response:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 数据载荷
      msg:
        description: 中文提示信息
        type: string
    type: object
...
  internal_transport_http.apiKeyResponse:
    properties:
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      key:
        type: string
      lastUsedAt:
        type: string
      name:
        type: string
    type: object
...
  internal_transport_http.authResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/internal_transport_http.userResponse'
    type: object
...
  internal_transport_http.createAPIKeyRequest:
    properties:
      expiresIn:
        description: 过期时间（如 "720h" 表示30天）
        type: string
      name:
        description: API Key名称/描述
        type: string
    required:
    - name
    type: object
...
  internal_transport_http.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  internal_transport_http.refreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  internal_transport_http.registerRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    type: object
  internal_transport_http.userResponse:
    properties:
      email:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      isEmailVerified:
        type: boolean
      tier:
        type: string
      username:
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: TempMail 后端 API 文档
  title: TempMail Backend API
  version: 0.9.0
paths:
...
  /v1/api-keys:
    get:
      description: 获取当前用户的所有API Key
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              count:
                type: integer
              items:
                items:
                  $ref: '#/definitions/internal_transport_http.apiKeyResponse'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 获取API Key列表
      tags:
      - APIKeys
    post:
      consumes:
      - application/json
      description: 为当前用户创建一个新的API Key用于第三方对接
      parameters:
      - description: API Key参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.createAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_http.apiKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 创建API Key
      tags:
      - APIKeys
  /v1/api-keys/{id}:
    delete:
      description: 删除指定的API Key
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 删除API Key
      tags:
      - APIKeys
    get:
      description: 获取指定API Key的详细信息
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http.apiKeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 获取API Key详情
      tags:
      - APIKeys
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: 使用邮箱和密码进行身份验证，成功后返回认证令牌
      parameters:
      - description: 登录凭证
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/internal_transport_http.authResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "401":
          description: 邮箱或密码错误
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "403":
          description: 账户已被禁用
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 用户登录
      tags:
      - 认证
  /v1/auth/me:
    get:
      description: 获取已认证用户的详细信息，需要有效的访问令牌
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/internal_transport_http.userResponse'
        "401":
          description: 未认证或令牌无效
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌，避免重新登录
      parameters:
      - description: 包含刷新令牌的请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 新的访问令牌
          schema:
            properties:
              accessToken:
                type: string
              expiresIn:
                type: integer
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "401":
          description: 刷新令牌无效或已过期
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 刷新访问令牌
      tags:
      - 认证
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账户，返回用户信息和认证令牌
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: '#/definitions/internal_transport_http.authResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "409":
          description: 邮箱已存在
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/internal_transport_http.Response'
      summary: 用户注册
      tags:
      - 认证
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 兼容层 API Key
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 使用格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
  MailboxToken:
    description: 邮箱访问令牌
    in: header
    name: X-Mailbox-Token
    type: apiKey
swagger: "2.0"